+++ src/byob.js
@@ -111,7 +111,7 @@ ArgLabelMorph*/
 
 // Global stuff ////////////////////////////////////////////////////////
 
-modules.byob = '2022-May-19';
+modules.byob = '2022-May-20';
 
 // Declarations
 
@@ -2675,6 +2675,12 @@ BlockEditorMorph.prototype.refreshAllBlockInstances = function (oldSpec) {
     } else {
         this.target.allDependentInvocationsOf(oldSpec).reverse().forEach(
             block => block.refresh(def)
+        );
+        this.target.parentThatIsA(StageMorph).allContextsInvoking(
+            def.blockSpec(),
+            this.target
+        ).forEach(
+            context => context.changed()
         );
     }
     if (template) {